app/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-Login==0.6.3

# Google Cloud dependencies
google-cloud-aiplatform==1.66.0
google-cloud-storage==2.10.0
google-cloud-logging==3.11.2
google-auth==2.34.0

# AI and Machine Learning
keras==3.5.0
google-ai-generativelanguage

# Utilities
requests==2.32.3
numpy==1.26.4
pillow==10.4.0
gunicorn==23.0.0

# Development Tools
flake8==3.8.4
isort==5.13.2

# Miscellaneous dependencies
cryptography==43.0.1
